pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "UsrAK"
version = "0.1.0"
authors = [
    {name = "frankmurrey"},
    {name = "preposition17"},
]
description = "A reusable FastAPI authentication and authorization extension."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: Session",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.24",
    "pydantic[email]>=2.11.0",
    "pyjwt>=2.10.0",
    "cryptography>=45.0.0",
    "passlib>=1.7.0",
    "tenacity>=9.1.0",
    "lmdb>=1.6.2",
    "psycopg2>=2.9.7",
    "bcrypt==4.0.1"
]

[project.urls]
"Homepage" = "https://github.com/baryber/usrak"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio==1.0.0",
    "httpx>=0.28.0",
    "redis>=6.2.0",
    "fakeredis>=2.29.0",
    "uvicorn>=0.34.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["usrak*"]

[tool.pytest]
strict_markers = true
python_files = "tests/test_*.py"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "docker_required: marks tests as requiring running docker containers",
]
testspaths = [
    "tests",
]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py310"
respect-gitignore = true

select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade (modern python)
    "B",    # bugbear (ловит баги)
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
]

ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]
show_error_codes = true
warn_unused_configs = true
ignore_missing_imports = true